Bug fixes for initial spotify-hydra commit.
Defined a hydra dict so it's easier to build in a custom one later.

Messed around with the hydra docstring way too long until I figured out that
':hint none' has a terrible error if you mess up and ask for a docstr of
a head that has no docstr. Fixed hydra shortcuts to remove a dupe and fix that
error.

Removed unused arg from spotify--hydra-status-format

Removed quotes from bound-and-true-p checks.