Reworking Spotify Connect API receiver side some more.
...checking in because there's a bug somewhere that just freezes my emacs when I
try to load spotify.el so... yay to that.

spotify-api-json.el is for dealing with received json from spotify-apy.el calls
to spotify connect api endpoints.

Tried to make it data-centric so that new fields can be more easily added and
`gethash'es can be reigned in to one file, maybe.

Changed spotify-connect to use spotify-api-json.

Moved some funcs from spotify-json to spotify-api-json.