Changed smudge cache values to alist, added test.
Changed name of `smudge-cache--status` to `smudge-cache--data`, since it
stores status, volume, & possibly more in future...

Changed `smudge-cache--data` to be an alist of alists, instead of alist
of hash-tables. Don't really have the volume of data required to make
hash-tables a requirement...

Moved timestamp functions together.

Tried to rename the `smudge-cache--data` functions to make more sense.
  - We'll see... "data", "device", "get", "set" are all over the place
    in the cache so it's tricky.

Added test for `smudge-cache--get-data` function.