Added unit test, fixed bugs.
Updated `test-smudge-cache--force-volume` to be able to force all, one,
or the other volumes in the cache.

Added unit test for `smudge-cache-get-volume` that tests that the volume
returned is the most recent one in the cache (not just blindly returning
`:volume` volue).

Fixed bug in `smudge-cache-get-volume` where one-invalid/one-valid
timestamp cases were improperly checked and so always used.