Added regression test.
`test-smudge-cache-update-status--nil-status` added.

There was a bug where after Spotify being paused for a while, the player
status update polling was receiving null player statuses, which raised
an error in `smudge-cache-update-status`.

The bug in `smudge-cache-update-status` was fixed in an earlier commit;
this is a regression test to make sure it stays fixed.